delete post (#1131)
