+++ apps/1CD/instagram/frontend/src/graphql/story.graphql
@@ -1,24 +1,7 @@
 mutation CreateStory($input: StoryInput!) {
   createStory(input: $input) {
     _id
-    userId {
-      _id
-      userName
-      fullName
-      email
-      phone
-      bio
-      gender
-      profileImg
-      accountVisibility
-      followerCount
-      followingCount
-      createdAt
-      updatedAt
-      otp
-      resetPasswordToken
-      resetPasswordTokenExpire
-    }
+    userId
     description
     image
     createdAt
